builtin support to send multipart HTML emails
multipart is enabled by default, because it is unlikely that sending
html mails excludes sending multipart ones, which are preferable.

I think it's important to say that this is NOT the best implementation
possible.
A better implementation would generate the text/plain part directly from
the feed, taking advantage of the different representations that could
be presented for a single entry.
This is, still, a good improvement for many users.

refs #101